requirements.txt
streamlit
marker-pdf
transformers
torch
pdf2image
numpy
pandas
Pillow
aiohttp